Relocating color to pojo package. Removing ko from board
